Refactor plan reading to separate module
